refactor: add Cart button component
