+++ src/components/product/product.tsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import { ROUTES } from "../../constants/routes";
-import { useCart } from "../../contexts/cart/cartContext";
 import { IProduct } from "../../interfaces/products";
+import ProductCartButton from "../ProductCartButton/ProductCartButton";
 import './Product.scss';
 
 interface IProductProps {
@@ -9,18 +9,6 @@ interface IProductProps {
 }
 
 export default function Product({ product }: IProductProps) {
-  const { cart, addToCart, removeFromCart } = useCart();
-
-  const inCart = cart.find(el => el.id === product.id);
-
-  const addHandler = () => {
-    addToCart(product);
-  }
-
-  const removeHandler = () => {
-    removeFromCart(product);
-  }
-
   return (
     <div className="product-item">
       <div className="item-wrapper" style={{background: `url(${product.thumbnail}) 0% 0% / cover`}}>
@@ -38,9 +26,10 @@ export default function Product({ product }: IProductProps) {
           </div>
         </div>
         <div className="item-buttons">
-          {inCart && <button className="button" onClick={removeHandler}>DROP FROM CART</button>}
-          {!inCart && <button className="button" onClick={addHandler}>ADD TO CART</button>}
-          <Link to={`${ROUTES.PRODUCT_DETAILS}/${product.id}`}><button className="button">DETAILS</button></Link>
+          <ProductCartButton product={product}/>
+          <Link to={`${ROUTES.PRODUCT_DETAILS}/${product.id}`}>
+            <button className="button">DETAILS</button>
+          </Link>
         </div>
       </div>
     </div>
